refactor: 지역선택 모달 컴포넌트 분리 및 타입 파일 분리
