+++ src/components/SelectRegion/index.js
@@ -1,0 +1,3 @@
+export { default as SidoTable } from './SidoTable';
+export { default as SigunguTable } from './SigunguTable';
+export { default as EupmyeondongTable } from './EupmyeondongTable';
